# src/public/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "API-TYPEORM.",
		"description": "Esta API foi criada com o intuito de aprimorar minhas habilidades em TypeScript, TypeORM, migrations, seeds e deploy, com o CRUD completas. Repositório [https://github.com/franciley45/API-typescript-typeORM](https://github.com/franciley45/API-typescript-typeORM)",
		"contact": {
			"email": "testepa45@gmail.com"
		},
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "https://api-typescript-type-orm-franciley45.vercel.app",
			"description": "API de Produção"
		},
		{
			"url": "http://localhost:3000",
			"description": "API de localhost"
		}
	],
	"paths": {
		"/users/login": {
			"post": {
				"summary": "Login de Usuário",
				"description": "Essa rota será responsável por Login do usuário",
				"tags": [
					"User"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/login"
							},
							"examples": {
								"user": {
									"value": {
										"email": "franciley@email.com",
										"password": "12345"
									}
								}
							}
						}
					}
				},
				"responses": {
					"400": {
						"description": "fill in all fields!"
					},
					"403": {
						"description": "unregistered user!"
					},
					"401": {
						"description": "wrong password!"
					},
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user_return_post"
								}
							}
						}
					}
				}
			}
		},
		"/users": {
			"post": {
				"summary": "Cadastro de Usuário",
				"description": "Essa rota será responsável por cadastrar um novo usuário",
				"tags": [
					"User"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/user"
							},
							"examples": {
								"user": {
									"value": {
										"name": "franciley rodrigues",
										"email": "test@gmail.com",
										"password": "12345"
									}
								}
							}
						}
					}
				},
				"responses": {
					"400": {
						"description": "fill in all fields!"
					},
					"409": {
						"description": "User already registered!"
					},
					"201": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user_return_post"
								}
							}
						}
					}
				}
			},
			"get": {
				"summary": "Busca todos os usuários.",
				"description": "Essa rota será responsável por busca todos os usuários.",
				"tags": [
					"User"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/user"
									}
								}
							}
						}
					}
				}
			}
		},
		"/user/{id}": {
			"get": {
				"summary": "Busca Usuário por Id",
				"description": "Essa rota será responsável por Busca Usuário por Id",
				"tags": [
					"User"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Id do Usuário para busca",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"404": {
						"description": "User not registered!"
					},
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user"
								}
							}
						}
					}
				}
			}
		},
		"/post": {
			"get": {
				"summary": "Busca todos os Posts.",
				"description": "Essa rota será responsável por busca todos os Posts",
				"tags": [
					"Post"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/post_milagre"
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Cria um novo Post ",
				"description": "Essa rota será responsável por Cria um novo.",
				"tags": [
					"Post"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/post_milagre"
							},
							"examples": {
								"Post": {
									"value": {
										"title": "A salvação esta em Jesus Cristo ",
										"date": "2023-09-21T09:02:55.000Z",
										"milagres": "salvação é perdão"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/post_milagre"
								}
							}
						}
					},
					"401": {
						"description": "Expired or invalid token!"
					},
					"400": {
						"description": "Token not found!"
					}
				}
			}
		},
		"/post/{id}": {
			"get": {
				"summary": "Busca Post por Id",
				"description": "Essa rota será responsável por Busca Post por Id",
				"tags": [
					"Post"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Id do Post para busca",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"404": {
						"description": "Post not registered!"
					},
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/post_milagre"
								}
							}
						}
					}
				}
			},
			"delete": {
				"summary": "Deleta Post por Id",
				"description": "Essa rota será responsável por Deleta Post por Id.",
				"tags": [
					"Post"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Id do Post para busca",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Expired or invalid token!"
					},
					"400": {
						"description": "Token not found!"
					},
					"404": {
						"description": "Post not found!"
					},
					"402": {
						"description": "this post is not yours!"
					}
				}
			},
			"put": {
				"summary": "Atualizar Post por Id",
				"description": "Essa rota será responsável por Atualizar Post por Id.",
				"tags": [
					"Post"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/post_milagre"
							},
							"examples": {
								"Post": {
									"value": {
										"title": "deus seja louvado por toda minha vida",
										"date": "2023-09-21T09:02:55.000Z",
										"milagres": "salvação é perdão"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Id do Post para busca",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/post_milagre"
								}
							}
						}
					},
					"401": {
						"description": "Expired or invalid token!"
					},
					"400": {
						"description": "Token not found!"
					},
					"404": {
						"description": "Post not found!"
					},
					"402": {
						"description": "this post is not yours!"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"user": {
				"type": "object",
				"properties": {
					"user_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				}
			},
			"user_return_post": {
				"type": "object",
				"properties": {
					"user_id": {
						"type": "number"
					},
					"name": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				}
			},
			"login": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				}
			},
			"post_milagre": {
				"type": "object",
				"properties": {
					"posts_id": {
						"type": "number"
					},
					"title": {
						"type": "string"
					},
					"user_id": {
						"type": "number"
					},
					"date": {
						"type": "string"
					},
					"milagres": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}